requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-ollama==0.1.0
langgraph==0.1.0
faiss-cpu==1.8.0
unstructured==0.14.0
python-dotenv==1.0.1